rename routes to demos
