+++ app/demos/covid/App.js
