+++ app/demos/covid/DataLoader.js
