+++ app/demos/covid/README.md
